# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Search Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UpdateSearchBody {
  title: string;
}

export type CreateSearchBodyMode =
  (typeof CreateSearchBodyMode)[keyof typeof CreateSearchBodyMode];

export const CreateSearchBodyMode = {
  quick: "quick",
  deep: "deep",
  expert: "expert",
} as const;

export interface CreateSearchBody {
  query: string;
  mode?: CreateSearchBodyMode;
}

export interface FollowUpBody {
  question: string;
}

export interface SearchSource {
  title: string;
  url: string;
  snippet: string;
  credibilityScore: number;
  domain: string;
  publishedDate?: string | null;
}

export interface SearchResult {
  id: string;
  query: string;
  mode: string;
  synthesis: string;
  sources: SearchSource[];
  followUps: string[];
  relatedQueries: string[];
  createdAt: string;
  /** Search duration in milliseconds */
  duration: number;
}

export interface SearchHistoryItem {
  id: string;
  query: string;
  title?: string | null;
  mode: string;
  sourceCount: number;
  createdAt: string;
  previewText: string;
}

export interface SearchHistoryList {
  items: SearchHistoryItem[];
  total: number;
}

export type TrendingTopicsTopicsItem = {
  topic: string;
  category: string;
  searchCount: number;
};

export interface TrendingTopics {
  topics: TrendingTopicsTopicsItem[];
}

export type SearchStatsTopModesItem = {
  mode: string;
  count: number;
};

export interface SearchStats {
  totalSearches: number;
  totalSources: number;
  avgSourcesPerSearch: number;
  avgDuration: number;
  topModes: SearchStatsTopModesItem[];
}

export type ListSearchHistoryParams = {
  limit?: number;
  offset?: number;
};