# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Search Agent API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Initiates a deep search — returns SSE stream with AI synthesis
 * @summary Create a new search (streaming SSE)
 */
export const CreateSearchBody = zod.object({
  query: zod.string(),
  mode: zod.enum(["quick", "deep", "expert"]).optional(),
});

/**
 * @summary Get recent search history
 */
export const ListSearchHistoryQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListSearchHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      query: zod.string(),
      title: zod.string().nullish(),
      mode: zod.string(),
      sourceCount: zod.number(),
      createdAt: zod.string(),
      previewText: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get trending search topics
 */
export const GetTrendingTopicsResponse = zod.object({
  topics: zod.array(
    zod.object({
      topic: zod.string(),
      category: zod.string(),
      searchCount: zod.number(),
    }),
  ),
});

/**
 * @summary Get search statistics overview
 */
export const GetSearchStatsResponse = zod.object({
  totalSearches: zod.number(),
  totalSources: zod.number(),
  avgSourcesPerSearch: zod.number(),
  avgDuration: zod.number(),
  topModes: zod.array(
    zod.object({
      mode: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get a specific search result
 */
export const GetSearchParams = zod.object({
  id: zod.coerce.string(),
});

export const GetSearchResponse = zod.object({
  id: zod.string(),
  query: zod.string(),
  mode: zod.string(),
  synthesis: zod.string(),
  sources: zod.array(
    zod.object({
      title: zod.string(),
      url: zod.string(),
      snippet: zod.string(),
      credibilityScore: zod.number(),
      domain: zod.string(),
      publishedDate: zod.string().nullish(),
    }),
  ),
  followUps: zod.array(zod.string()),
  relatedQueries: zod.array(zod.string()),
  createdAt: zod.string(),
  duration: zod.number().describe("Search duration in milliseconds"),
});

/**
 * @summary Delete a search from history
 */
export const DeleteSearchParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Rename a search (update its display title)
 */
export const UpdateSearchParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateSearchBody = zod.object({
  title: zod.string(),
});

export const UpdateSearchResponse = zod.object({
  id: zod.string(),
  query: zod.string(),
  title: zod.string().nullish(),
  mode: zod.string(),
  sourceCount: zod.number(),
  createdAt: zod.string(),
  previewText: zod.string(),
});

/**
 * @summary Ask a follow-up question on an existing search (streaming SSE)
 */
export const CreateFollowUpParams = zod.object({
  id: zod.coerce.string(),
});

export const CreateFollowUpBody = zod.object({
  question: zod.string(),
});