# lib/api-zod/src/generated/types/searchHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Search Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchHistoryItem {
  id: string;
  query: string;
  title?: string | null;
  mode: string;
  sourceCount: number;
  createdAt: string;
  previewText: string;
}